% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadraticFormToMatrix.R
\name{quadraticFormToMatrix}
\alias{quadraticFormToMatrix}
\title{Transformation of the quadratic conic representation into the matrix representation.}
\usage{
quadraticFormToMatrix(v)
}
\arguments{
\item{v}{a \eqn{(6 \times 1)} vector of the parameters \eqn{(a, b, c, d, e, f)} 
of the quadratic form \eqn{ax^2 + bxy + cy^2 + dxz + eyz + fz^2 = 0}.}
}
\value{
A \eqn{(3 \times 3)} matrix representation of the conic (symmetric matrix).
}
\description{
Transformation of the quadratic conic representation into the matrix representation.
}
\examples{
v <- c(2,2,-2,-20,20,10)
quadraticFormToMatrix(v)
}
