% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjoint.R
\name{adjoint}
\alias{adjoint}
\title{Adjoint matrix}
\usage{
adjoint(A)
}
\arguments{
\item{A}{a square matrix.}
}
\value{
The adjoint matrix of A (square matrix with the same dimension as A).
}
\description{
Compute the classical adjoint (also called adjugate) of a square matrix. The adjoint is the transpose of the cofactor matrix.
}
\examples{
A <- matrix(c(1,4,5,3,7,2,2,8,3),nrow=3,ncol=3)
A
B <- adjoint(A)
B
}
\seealso{
\code{\link{cofactor}}, \code{\link{minor}}
}
